Move `LiquidityManager` to `manager.rs`
.. and `LSPS0MessageHandler` to `lsps0/message_handler.rs`
