+++ src/lsps0/mod.rs
@@ -11,4 +11,3 @@
 
 pub mod message_handler;
 pub mod msgs;
-pub mod protocol;
